+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * Added `nsim` argument to `vi_permute()` for reducing the sampling variability induced by permuting each predictor [(#36)](https://github.com/koalaverse/vip/issues/36).
 
+* Added `sample_size` and `sample_frac` arguments to `vi_permute()` for reducing the size of the training sample for every Monte Carlo repition [(#41)](https://github.com/koalaverse/vip/issues/41).
+
 * Greatly improved the documentation for `vi_model()` and the various objects it supports.
 
 * New argument `rank`, which default to `FALSE`, available in `vi()` [(#55)](https://github.com/koalaverse/vip/issues/55).